+++ NEWS.md
@@ -2,6 +2,9 @@
 * Can now `deriveArray()` using expressions to create the subvariables.
 * New function `slideCategories()` helps you create overlapping categorical 
 variables (#396).
+* crunch will work with upcoming changes to `stringsAsFactors` defaults (#402).
+* Exporting a Deck requires (and now includes) a valid JSON body in its request.
+* Documentation for `newSlide` includes examples of `vizType` settings and other improvements.
 
 # crunch 1.26.2
 * A new `copyFolders()` function that copies folders and variable order from one dataset to another (similar to `copyOrder()` which was deprecated)